make sure k and a are of type float

+++ pyccl/halos/halo_model.py
@@ -382,8 +382,8 @@ def halomod_mean_profile_1pt(cosmo, hmc, k, a, prof,
         `k` and `a` respectively. If `k` or `a` are scalars, the
         corresponding dimension will be squeezed out on output.
     """
-    a_use = np.atleast_1d(a)
-    k_use = np.atleast_1d(k)
+    a_use = np.atleast_1d(a).astype(float)
+    k_use = np.atleast_1d(k).astype(float)
 
     # Check inputs
     if not isinstance(prof, HaloProfile):
@@ -436,8 +436,8 @@ def halomod_bias_1pt(cosmo, hmc, k, a, prof, normprof=False):
         `k` and `a` respectively. If `k` or `a` are scalars, the
         corresponding dimension will be squeezed out on output.
     """
-    a_use = np.atleast_1d(a)
-    k_use = np.atleast_1d(k)
+    a_use = np.atleast_1d(a).astype(float)
+    k_use = np.atleast_1d(k).astype(float)
 
     # Check inputs
     if not isinstance(prof, HaloProfile):
@@ -530,8 +530,8 @@ def halomod_power_spectrum(cosmo, hmc, k, a, prof,
         `k` and `a` respectively. If `k` or `a` are scalars, the
         corresponding dimension will be squeezed out on output.
     """
-    a_use = np.atleast_1d(a)
-    k_use = np.atleast_1d(k)
+    a_use = np.atleast_1d(a).astype(float)
+    k_use = np.atleast_1d(k).astype(float)
 
     # Check inputs
     if not isinstance(prof, HaloProfile):
@@ -788,8 +788,8 @@ def halomod_trispectrum_1h(cosmo, hmc, k, a,
         If `k` or `a` are scalars, the corresponding dimension will
         be squeezed out on output.
     """
-    a_use = np.atleast_1d(a)
-    k_use = np.atleast_1d(k)
+    a_use = np.atleast_1d(a).astype(float)
+    k_use = np.atleast_1d(k).astype(float)
 
     # Check inputs
     if not isinstance(prof1, HaloProfile):
